requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-flex-fields==1.0.2
drf-writable-nested==0.7.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
mssql-django==1.3
mysqlclient==2.2.1
packaging==23.2
pillow==10.2.0
pyodbc==5.0.1
pyTelegramBotAPI==4.14.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
telebot==0.0.5
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
